requirements.txt
discord.py
python-dotenv
mcrcon
mcstatus